[Feat] set text styles
