+++ src/app/blog/[category]/page.tsx
@@ -5,6 +5,7 @@ type Props = {
     category: string;
   };
 };
+
 function TitleList(category: string) {
   const fs = require("fs");
   const path = `${process.cwd()}/src/posts/${category}`;
@@ -15,10 +16,15 @@ const Category = async ({ params }: Props) => {
   // const post = await getPostDetail(params.category, params.id);
   return (
     <div>
-      <div className={"text-5xl"}>{params.category}</div>
+      <div className={"m-5 text-5xl"}>{params.category}</div>
       {TitleList(params.category).map((title: string, idx: any) => (
         <div key={idx}>
-          <Link href={`/blog/${params.category}/${title}`}>{title}</Link>
+          <Link
+            href={`/blog/${params.category}/${title}`}
+            className={"m-5 text-2xl hover:text-amber-700"}
+          >
+            {title}
+          </Link>
         </div>
       ))}
     </div>
