+++ src/components/example.ts
@@ -1,0 +1,10 @@
+async function getData() {
+  const res = await fetch(`${process.env.API_URL}/api`, {
+    cache: "no-store",
+  });
+  if (!res.ok) {
+    // This will activate the closest `error.js` Error Boundary
+    throw new Error("Failed to fetch data");
+  }
+  return res.json();
+}
