+++ src/config/globals.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 @layer components {
-  .btn {
+  .topbar {
     @apply font-bold text-3xl py-2 px-4 border-b-2;
   }
 }
